requirements.txt
flask==2.3.3
openai==1.82.0
python-dotenv==1.0.0
exa-py==1.0.9
firecrawl-py==2.7.1
web3==7.10.0
requests==2.32.3 